Splitting healpix into 2. Adding adjacency matrix.
healpix_sel.py now healpix.py and healpix_draw.py, which is a drawing
playground for my own purposes.

Some cleanup for cart2.py: making exceptions a bit less fragile and a
bit more informative, and other small cleanups, all are cosmetic only.

Adding an adjacency matrix class, I was pretty excited about this, but
it seems to have truly bad performance capabilities, so it'll probably
have to be ditched.